admin_api/lambda_functions/requirements.txt
python-decouple
boto3
requests
mysql-connector-python
regex
